+++ Assets/Scripts/Note_Scripts/Note_SOs/Note_E_Attack.cs
@@ -5,14 +5,14 @@ using UnityEngine;
 [CreateAssetMenu(menuName = "Notes/Attack")]
 public class Note_E_Attack : Note_SO
 {
-    [SerializeField] private float damage;
-    [SerializeField] private float upgradeDamage;
+    [SerializeField] private int damage;
+    [SerializeField] private int upgradeDamage;
     public override void Apply(Note_Effect_Handler noteEffectHandler, Note_SO note)
     {
         noteEffectHandler.DealDamage(note, GetTotalStat(Game_Manager.instance.statsManager.GetStackCount(note)));
     }
 
-    public override float GetTotalStat(int ownedCount)
+    public override int GetTotalStat(int ownedCount)
     {
         return damage + upgradeDamage * (ownedCount - 1);
     }
